+++ src/scoring/calcLightness.ts
@@ -3,7 +3,7 @@ import { getContrastScoreForObjective } from './getContrastScoreForObjective'
 import type { ContrastConfig_PREPARED } from '../contrast/contrastConfig'
 import { ColorSpace } from '../types'
 import { chromaLimits } from '../utils/chromaLimits'
-import { clampColorToSpace } from '../utils/clampColorToSpace'
+import { clampColorToSpace } from '../clamp/clampColorToSpace'
 import { log } from '../utils/log'
 import { signOf } from '../utils/misc'
 import { lightnessAndPatch } from '../light/lightnessAndPatch'
